+++ infra/docker/postgres/init/001-create-test-database.sql
@@ -1,0 +1,4 @@
+-- SPDX-FileCopyrightText: 2026 OpenDX CompanyOS contributors
+-- SPDX-License-Identifier: Apache-2.0
+
+CREATE DATABASE opendx_test;